Added the code of in-class exercises
InClassExe1-MonocolorLed: includes both Arduino code and Android code and demonstrate the control of on/off on a LED.

InClassExe2-RGBLed: includes both Arduino code and Android code and demostrates that changing the RGB LED's brightness by sliding the Seekbar in the Android app.

InClassExe3-BiDirectonalControl: includes both Arduino code and Android code and demonstrates that changing the LED's brightness by either sliding the Seekbar on the Android app or turning the potentiometer wired to the board. The Seekbar also reflects the brightness of the LED accordingly.